# lib/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Repody",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/workflows": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "List Workflows",
        "operationId": "list_workflows_v1_workflows_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Create Workflow",
        "operationId": "create_workflow_v1_workflows_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkflowBody"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/bulk-delete": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Bulk Delete Workflows",
        "operationId": "bulk_delete_workflows_v1_workflows_bulk_delete_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteWorkflowsBody"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}": {
      "get": {
        "tags": [
          "workflows"
        ],
        "summary": "Get Workflow",
        "operationId": "get_workflow_v1_workflows__workflow_id__get",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "workflows"
        ],
        "summary": "Update Workflow",
        "operationId": "update_workflow_v1_workflows__workflow_id__put",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WorkflowSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "workflows"
        ],
        "summary": "Delete Workflow",
        "operationId": "delete_workflow_v1_workflows__workflow_id__delete",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/deploy": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Deploy Workflow",
        "operationId": "deploy_workflow_v1_workflows__workflow_id__deploy_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/DeployWorkflowBody"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/workflows/{workflow_id}/dry-run": {
      "post": {
        "tags": [
          "workflows"
        ],
        "summary": "Dry Run",
        "operationId": "dry_run_v1_workflows__workflow_id__dry_run_post",
        "parameters": [
          {
            "name": "workflow_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Workflow Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DryRunBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DryRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BulkDeleteWorkflowsBody": {
        "properties": {
          "ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Ids"
          }
        },
        "type": "object",
        "required": [
          "ids"
        ],
        "title": "BulkDeleteWorkflowsBody"
      },
      "CallSeriesPointSchema": {
        "properties": {
          "day": {
            "type": "string",
            "title": "Day"
          },
          "calls": {
            "type": "integer",
            "title": "Calls"
          }
        },
        "type": "object",
        "required": [
          "day",
          "calls"
        ],
        "title": "CallSeriesPointSchema"
      },
...
      "CreateWorkflowBody": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name",
            "default": "Untitled workflow"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": "Me"
          }
        },
        "type": "object",
        "title": "CreateWorkflowBody"
      },
      "DeployWorkflowBody": {
        "properties": {
          "apiKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey"
          }
        },
        "type": "object",
        "title": "DeployWorkflowBody"
      },
      "DocumentDefSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "documentType": {
            "type": "string",
            "title": "Documenttype",
            "default": ""
          },
          "extractionMode": {
            "type": "string",
            "title": "Extractionmode",
            "description": "Read path id from GET /processing-paths (document_model).",
            "default": "document_model"
          },
          "validationMode": {
            "type": "string",
            "title": "Validationmode",
            "description": "logic_only",
            "default": "logic_only"
          },
          "ocrModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocrmodel",
            "default": "repody:vlm"
          },
          "schema": {
            "items": {
              "$ref": "#/components/schemas/SchemaFieldSchema"
            },
            "type": "array",
            "title": "Schema"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "DocumentDefSchema"
      },
...
      "DryRunBody": {
        "properties": {
          "fields": {
            "items": {
              "$ref": "#/components/schemas/DryRunFieldInput"
            },
            "type": "array",
            "title": "Fields",
            "default": []
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/DryRunRuleInput"
            },
            "type": "array",
            "title": "Rules",
            "default": []
          },
          "documents": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/DocumentDefSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Documents"
          },
          "rulesFull": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/WorkflowRuleSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rulesfull"
          }
        },
        "type": "object",
        "title": "DryRunBody"
      },
      "DryRunExtracted": {
        "properties": {
          "field": {
            "type": "string",
            "title": "Field"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "matched": {
            "type": "boolean",
            "title": "Matched"
          }
        },
        "type": "object",
        "required": [
          "field",
          "value",
          "matched"
        ],
        "title": "DryRunExtracted"
      },
      "DryRunFieldInput": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "sampleValue": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Samplevalue"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "DryRunFieldInput"
      },
      "DryRunResponse": {
        "properties": {
          "extracted": {
            "items": {
              "$ref": "#/components/schemas/DryRunExtracted"
            },
            "type": "array",
            "title": "Extracted"
          },
          "ruleResults": {
            "items": {
              "$ref": "#/components/schemas/DryRunRuleResult"
            },
            "type": "array",
            "title": "Ruleresults"
          }
        },
        "type": "object",
        "required": [
          "extracted",
          "ruleResults"
        ],
        "title": "DryRunResponse"
      },
      "DryRunRuleInput": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "reject"
          },
          "conditions": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "conditionJunction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditionjunction"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "title": "DryRunRuleInput"
      },
      "DryRunRuleResult": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "kind",
          "status",
          "detail"
        ],
        "title": "DryRunRuleResult"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SchemaFieldSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "SchemaFieldSchema"
      },
...
      "TopFailingRuleSchema": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          }
        },
        "type": "object",
        "required": [
          "name",
          "count",
          "severity"
        ],
        "title": "TopFailingRuleSchema"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkflowApiStatsSchema": {
        "properties": {
          "apiCallsToday": {
            "type": "integer",
            "title": "Apicallstoday"
          },
          "apiCallsTotal": {
            "type": "integer",
            "title": "Apicallstotal"
          },
          "avgLatencyMs": {
            "type": "integer",
            "title": "Avglatencyms"
          },
          "callSeries": {
            "items": {
              "$ref": "#/components/schemas/CallSeriesPointSchema"
            },
            "type": "array",
            "title": "Callseries"
          },
          "topFailingRules": {
            "items": {
              "$ref": "#/components/schemas/TopFailingRuleSchema"
            },
            "type": "array",
            "title": "Topfailingrules"
          }
        },
        "type": "object",
        "required": [
          "apiCallsToday",
          "apiCallsTotal",
          "avgLatencyMs",
          "callSeries",
          "topFailingRules"
        ],
        "title": "WorkflowApiStatsSchema"
      },
      "WorkflowListResponse": {
        "properties": {
          "workflows": {
            "items": {
              "$ref": "#/components/schemas/WorkflowSchema"
            },
            "type": "array",
            "title": "Workflows"
          }
        },
        "type": "object",
        "required": [
          "workflows"
        ],
        "title": "WorkflowListResponse"
      },
      "WorkflowResponse": {
        "properties": {
          "workflow": {
            "$ref": "#/components/schemas/WorkflowSchema"
          }
        },
        "type": "object",
        "required": [
          "workflow"
        ],
        "title": "WorkflowResponse"
      },
      "WorkflowRuleSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "default": "logic"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "default": "intra"
          },
          "appliesTo": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Appliesto",
            "default": []
          },
          "conditions": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditions"
          },
          "conditionJunction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conditionjunction"
          },
          "body": {
            "type": "string",
            "title": "Body",
            "default": ""
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "default": "reject"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "WorkflowRuleSchema"
      },
      "WorkflowSchema": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "default": ""
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "draft"
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": "Me"
          },
          "lastRun": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastrun"
          },
          "successRate": {
            "type": "number",
            "title": "Successrate",
            "default": 0.0
          },
          "totalRuns": {
            "type": "integer",
            "title": "Totalruns",
            "default": 0
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentDefSchema"
            },
            "type": "array",
            "title": "Documents",
            "default": []
          },
          "rules": {
            "items": {
              "$ref": "#/components/schemas/WorkflowRuleSchema"
            },
            "type": "array",
            "title": "Rules",
            "default": []
          },
          "deployedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deployedat"
          },
          "apiKey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikey",
            "description": "Plaintext key \u2014 returned only once on deploy."
          },
          "apiKeyHint": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apikeyhint",
            "description": "Masked hint for deployed workflows."
          },
          "defaultLlmModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Defaultllmmodel"
          },
          "apiStats": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkflowApiStatsSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "WorkflowSchema"
      }
    }
  }
}